[HW][ExportModuleHierarchy] Allow multiple output files per module (#2752)
From the FIRRTL flow we need to be able to specify multiple output files
for the module hierarchy.  To do this, this change switches the
`firrtl.moduleHierarchyFile` to have an array of files instead of a
single file.

The FIRRTL flow needs to be able to output a module hierarchy file for
the testharness (top level module) and the DUT module. When these two
modules are the same, we would end up only outputing a single hierarchy
file.  Unfortunately, the SiFive build system is really hard coded to
expect both files (for now) so we had to add the ability to generate the
same file twice.

Co-authored-by: Megan Wachs <megan@sifive.com>