[FIRRTL][Inliner] Fix NLA filtering on symbol collision
When a module is inlined into a parent module, the inlined body may
define symbols that have to be changed to not collide with existing
symbols in the parent module. When this happens, the code will find all
NLAs targeting referencing this symbol and update the symbol referenced.

When a module is inlined into two locations, it is important that the
NLAs referring to each location are updated with the renamed symbol at
that location.  There was a bug where we would update the symbol of all
NLAs to use the same symbol.  This would cause a verifier error after
the module InlinerPass:

```
'firrtl.nla' op  operation with symbol: #hw.innerNameRef<@ModuleName::@target> does not contain a reference to the NonLocalAnchor
```

This fixes the change by only updating the symbol in an NLA when the
current path matches the NLA's path.

One confusing part of this change is that InstanceOp symbol updates
still need to unconditionally update the non-leaf portion of an NLA's
path.  To implement this, if an annotation is a breadcrumb (has class
`circt.nonlocal`) we unconditionally rename the symbol in the referenced
NLA.