[FIRRTL][HW] Added module visibility attributes
`firrtl` and `hw` modules can now be annotated with MLIR visibility attributes.

The FIR parser sets the visibility of all non-main modules to private if the
main module is defined, otherwise all modules remain public.

Transformations were altered to inspect the visibility attribute instead of
checking whether a module matches the main of a circuit.