[FIRRTL][LowerToHW] Find the DUT in a prepass
There is an issue during module processing where we needed to know if
the module was instantiated underneath the DUT, and the answer to this
question depended on if we found the DUT module yet.

The typical way we have been handling this is to move this logic into a
post-pass, after we have found the DUT.  I don't think that this is the
first time this has burned us, so to make it safe we would need to
remove the DUT from the CircuitLoweringState to make sure no one used it
before it was populated.

Instead, this change finds the DUT in a pre-pass, so it is available
while processing each module.  This let me remove a post-pass to set the
testbench directory. We were already doing a pre-pass of the circuit to
look for NLAs, so I just stuck it in there.