[FIRRTL] Fix filelist filenames in BlackBoxReader
Fix a bug in BlackBoxReader where the filelist was not being correctly
populated.  This was adding the filename when this needed to set the
filename with the path (which may be different).

Cleanup the filelist by stripping any leading "./".  Also, strip
location information as this will cause problems if the output file is
ever read.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>