[FIRRTL] Enable "other" directory, BlackBoxReader
Add functionality to BlackBoxReader such that it will use the
"output_file" information on an external module when determining where
to put the black box.  The policy is as follows:

1. If the output file is a file (NOT a directory), then this file is
   where the black will written.
2. If the output file is a directory, then this becomes a lowest
   priority fallback for setting the directory.

This policy may require tweaking, but it gets CIRCT to match the Scala
FIRRTL Compiler (SFC) in tests and works for some SiFive designs which
are sensitive to the location of Grand Central black boxes.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>