[Moore] Add concat expr
As a first step towards bringing the Moore IRs into the CIRCT project,
add a `ConcatOp` to the Moore dialect which represents a concatenation
expression. The intention is to make this MLIR op precisely capture the
semantics specified in the SystemVerilog standard. The documentation for
the op points at the relevant pieces of the standard.

This op will allow Moore to start emitting `moore.mir.concat` instead of
directly doing the lowering to LLHD/Comb, which is a great starting
point to move the code generation logic from the Rust code into CIRCT.