[MSFT] [PyCDE] Make placement DB global rather than rooted
Since we are generating only pieces of a design (and FPGA resources are
obviously a property of the entire design), the devdb should be
independent of the 'top' module.

Also, since the devdb is supposed to act as an 'alternative view' of the
IR, load the IR's placements by default.