[FIRRTL] Infer type of mux with aggregate operands
Extend the type inference for the `MuxPrimOp` to also support operands
with aggregate types. The multiplexer is expected to return a widthless
type when the two operands are of an integer type but with unknown
widths, and the larger of the two otherwise (e.g., `UInt` for
`UInt, UInt<42>`). This inference only occurred for ground types and did
not recurse into vectors and bundles. This commit changes the type
inference to perform a pairwise inference on the leaves of arbitrarily
nested aggregates.

Fixes #2806.