[FIRRTL] Fix port symbol creation logic edge case in LowerTypes.
It was previously possible for LowerTypes to generate duplicate
symbols for different ports inside addArg. The logic that decides to
create the 'newSym' attribute previously checked whether the old
symbol was null, and if not, if it was empty. But the logic for
setting the 'sym' string in the first place was missing the empty
check. Because we create empty symbols by default in some cases, this
was leading to incorrect behavior, where the suffix was concatenated
to an empty string. If multiple lowered bundle types had the same
suffix, this could create duplicates.

Fortunately the solution is simple, we just need to include the empty
check in both places. To make this more clear, the logic is pulled out
into a shared variable that can be checked in both places.