[Support] Extract ValueMapper to separate support file (#2725)
This commit extracts the ValueMapper class from HandshakeToHW to a separate support file.

The ValueMapper facilitates the definition and connection of SSA def-use chains between two locations - a 'from' location (defining use-def chains) and a 'to' location (where new operations are created based on the 'from' location). An optional type transformation function may be provided to transform output types.
The mapper works for both SSACFG and GraphRegions, with the latter implemented through leveraging a BackedgeBuilder.