[HW] Support hw.instance of modules with parametric in/out types
To extend the capabilities of parametric modules, this commit introduces support for `hw.instance`'iating `hw.module`s with parametric input/output types. Upon instantiation (and `hw.instance` verification), module types are resolved based on the set of provided parameters to the `hw.instance` op.

The brunt of the work for supporting this is added through two new functions:
- resolveParametricType: Based on a set of parameters, a parametric type is resolved to its actual type. For now, only parametric `!hw.int` types are supported, but #2703 will add support for parametric array types.
- evaluateParamAttr: Based on a set of parameters, a parametric expression is evaluated through use of the existing ParamExprAttr canonicalizers.