[FIRRTL][Dedup] Delay fixup of connects (#2830)
When deduping modules, dedup ignores the names of fields in bundles.
After deduping two modules, the result type of an InstanceOp can change
when the field names of the bundle change. The connect op verifier
insists that the names of two connect bundles match, so we have to
"legalize connects" by walking instance results and potentially
decomposing a bulk connect to multiple connects.

There was a problem in dedup where it would "legalize connects"
immediately after deduping two modules. This would cause the body of the
parent module of the deduped module to change, which could block the
deduplication of the parent module.

This problem could be solved by a couple different options, such as
expanding connects in the parser or trying to make dedup smart enough to
see through these structural differences.

This PR opts to delay the "connect legalization" until after all module
deduping was done. Not only do we dedup more modules this way, but we
avoid a lot of intermediate legalization that can become unnecessary.

The code to fixup partial connects was deleted.