[FIRRTL] Relax reg reset width checking
The FIRRTL spec specifies that reset value of a register must be the
same width or smaller than the register.  We recently increased the
strictness of this check[0].  It turns out that SFC does not implement
this kind of error checking, and by not allowing implicit truncation we
are failing to build some cores which build fine on SFC.  This reverts
the strict width checking.