[HWMemSimImpl] Fix randomization of wide memories
The `HWMemSimImpl` pass currently generates SystemVerilog that some
tools cannot parse properly in cases where the memory data width is not
a multiple of the random value bit width (32). In those cases the
randomization produces:

    reg [33:0] _RANDOM;
    _RANDOM = {{`RANDOM}[1:0], {`RANDOM}};

Some tools do not recognize `{`RANDOM}[1:0] as a primary expression and
mistake it for being part of the concatenation, only to then complain
that this isn't proper concatenation syntax.

This commit changes the randomization code to put the part-select after
the overall concatenation, in the hope that most tools will have an
easier time identifying `{...}[1:0]` as a primary expression if it
appears as the right-hand side of an assignment. This now produces:

    reg [33:0] _RANDOM;
    _RANDOM = {{`RANDOM}, {`RANDOM}}[33:0];