[ExportVerilog] Remove max token limit and spilling in ExportVerilog. (#2802)
This is somewhat involved, and complicates ExportVerilog
significantly. Since we have developed a solution to handle this
problem in #2795, we can remove this code, and address this scenario
in PrepareForEmission.

It is also important to remove this code because it has bugs. I
believe they could have been resolved with #2773, but ultimately it is
better to do this in PrepareForEmission.

If the new logic doesn't cause a spill when the old logic would have,
the worst case is the expression remains inline. This might cause
Verilator to complain. But the previous logic could potentially drop
wires in illegal places, and it is best to remove that.