[FIRRTL][Dedup] Directly use SHA256 byte array instead of converting to string.
In a recent commit, 654d5e96c84c80274f04ddd1a5984adc547b5cec, that
bumped the LLVM submodule, LLVM changed the SHA256 hashers to return
std::array<uint8_t, 32> instead of StringRefs. We attempted to avoid
dealing with this in FIRRTL's Dedup pass by explicitly converting the
array back into a string, but the way the casts were being done was
causing issues with MSVC.

Changing the code to work directly with the std::array is more direct
and efficient, and obviates the issues with casting.

This adds a custom implementation for the DenseMapInfo trait
specifically for std::array<uint8_t, 32>, which allows us to directly
use the SHA256 hash results as map keys.