[FIRRTL][Dedup] enhance error messages for dedup failures (#2887)
A user can add an annotation "MustDedup" which lists modules which they
expect to dedup, and emits an error message when that fails. The error
message produced by a dedup failure doesn't give any hint as to why
deduplication failed. This can be very annoying to debug, and usually
involves diffing the FIRRTL IR modules.

This change adds a diagnostic comparison of two modules, looking for the
first difference between them which would have blocked deduplication.
This comparison ignores differences in port names and bundle element
names the same way that the structural hasher does.

This approach is still not super useful when trying to debug
deduplication failures: it will print the first difference found in the
modules and then bails out. This can help with simple things like when a
wire has a different type and you can easily map it back to the original
Chisel code. For differences in complicated logic, it can be challenging
to map it back to the chisel code and even harder to figure out why
these were emitted differently.