[HW] Add a mutable specialization of `SymbolCache` (#2919)
As proposed in #2882, it would be nice to have a `SymbolCache` which does not have a strict two-phase ordering to allow for mutation. The general usecases can be described as those where we, in a loop:
1. Perform a symbol cache lookup
2. Create a new module based on the results of that lookup
3. Update the symbol cache to reflect the existance of the new module

Without the ability to mutate the symbol cache, we'd need to rebuild it upon every loop iteration (defeating the purpose of having the cache).