[HW] Add a parametric `hw.module` instance specialization pass (#2882)
This commit introduces the `-hw-specialize` pass. Any `hw.instance` operation instantiating a parametric `hw.module` will trigger a specialization procedure which resolves all parametric types and values within the module based on the set of provided parameters to the `hw.instance` operation. This specialized module is created as a new `hw.module` and the referring `hw.instance` operation is rewritten to instantiate the newly specialized module.

The intention of this is to have a path for:
1. Writing generic modules ->
2. Specialize module instances ->
3. module inlining ->
4. inline optimization