[FIRRTL][ModuleInliner] Don't prefix memory names
A memory operation has a name attribute which is used for both the name
of the generated module and all instances of that module.  The
ModuleInliner pass is supposed to prefix the name of instance but not
the name of the module.  This change stops prefixing the memory, which
corrects the name of the module, but gives the incorrect instance name.
The assumption here is that its more important to get the name of the
memory module correct than the instance.