[FIRRTL] Delete GCT taps that have no ports
Fix a bug in the GrandCentralTaps pass where a data tap module that had
no ports would result in external module instantiations in the final
design that had no corresponding implementation.  Change the
GrandCentralTaps pass to delete these modules and their instantiations.
This can arise more frequently due to a recent change to the Grand
Central Chisel API where zero-width taps are now allowed in the API,
generate a tap module, but generate no actual taps.

An alternative implementation would be to generate empty modules which
are cleaned up later.  This seemed wasteful.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>