[CheckCombCycle] Improve error message
This commit adds `print-simple-cycle` option to print a simple cycle.
With that option, CheckCombCycle pass prints signal names contained
in a cycle (in the order of a cycle path) instead of dumping all
operations in the scc. It is useful for chisel users to debug their
chisel code since the cycle is actually a counterexample. Previous
error message (i.e. dumping all operations in SCC) is intentionally
kept to test the cycle dectection algorithm.