[FIRRTL][HW] Omitted unnecessary wires in LowerToHW
Instead of emitting a sv.wire which is eliminated later on by ExportVerilog,
wires are eagerly eliminated in `LowerToHW`.  Using a BackedgeBuilder,
placeholders are created until all values assigned by connects are collected
and resolved to a single op.

The `notEmitDuplicateWiresThatWereUnInlinedDueToLongNames` test in `hw-dialect`
is changed since `LowerToHW` and `prepareForEmission` now elimiante wires and
there is no name-length criteria to stop that.

Tested on large internal designs, in all instances fewer wires are emitted.