[FIRRTL] Add emitJSON attribute for GCT-SM
Add an attribute to the FIRRTL Circuit-level Grand Central (GCT) Signal
Mappings (SM) annotation to indicate if JSON should be emitted during
the GCT-SM pass.  This effectively encodes if firtool is compiling a
subcircuit (JSON needs to be emitted) vs. firtool is compiling the main
circuit (JSON should not be emitted).

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>