[FIRRTL] Do not dedup testharness memories (#2968)
This commit adds a unique groupID to each Testharness memory to prevent their
Dedup. This is a temporary fix to match the SFC behavior.
For any memory under the Testharness hierarchy, assign a unique groupID to it.
This pass uses the max unsigned integer to generate the groupIDs, the assumption
being that, it will never conflict with an existing groupID. The groupID will ensure
 that they are not deduped.

Note: Currently, the memory dedup happens during the metadata emission, because
 that is the first pass that requires the name of the final memory module. The plan
is to create a new Memory lowering pass, and move all the memory transformations to it.