Cmdline flags to deal with subcircuit (#2981)
Cmdline flags to deal with subcircuit.  subcircuits need to print xmrs relative to the DUT, which they don't know due to their separate compilation.  The toplevel module might also have been prefixed.  This is all less than ideal, but the mechanism is expected to be removed from Chisel soon, so this is a stopgap.