[FIRRTL][Dedup] Fix MustDedup crash
When two modules marked with MustDedup are not deduped with each other,
but they are deduped with something else, we end up trying to compare
the body of a module which was deleted.  This change makes us look up
the module they were actually deduped with to compare the modules.