[FIRRTL][ModuleInliner] Don't remove flatten annotation immediately
We need to keep it around to know to inline-flatten
should it be visited again (multiple sites, for example),
and to ensure flattening occurs when it is reached in the worklist.

This fixes TestInliningFlatten, but breaks annotation cleanup.

This partially reverts e1709ff50921a9c39d0e7a97c64ab903cf72dd9a.