[FIRRTL][ModuleInliner] Fix removing flatten annotations
The previous code removed it from the temporary AnnotationSet object,
not the module itself.  Also the code skipped this line when the
annotation was present (shouldFlatten was true -> continue).

Now the annotation is removed (only) when the module has been flattened.

It may still be visited for inlining on a separate path,
but since flattening has occurred just inlining as normal is fine.
(inlineInto instead of flattenInto).