[FIRRTL] print mem.conf even if there are no memories (#2991)
Currently, when passed `--repl-seq-mem-file=mem.conf` and there are no
memories, we will not emit the `mem.conf` metadata. This changes that
behaviour to create an empty file, which is the same as SFC behaviour.
Some tooling expects this file to be created even if it is empty.  We
made a similar change for `seq_mems.json` and previously decided we
didn't need to have the same behaviour for this file.