[SV] Avoid temporary wires when ReadInOutOp is after use (#3009)
When the ReadInOutOp accessing a register or a wire was after the use of the read value, a temporary wire was emitted.  This patch instead moves both the definition of the read operation and the read itself to the top of the module, ensuring the order is correct and no temporaries are needed.

As a result, the verilog representation of a large design is 0.6% shorter, with thousands of temporaries eliminated.