[FIRRTL][Dedup] Improve error when module is both NoDedup and MustDedup (#3032)
When a module is not deduplicated due to a NoDedup annotation, the
equivalence checker would compare the module bodies, missing the fact
that the annotation was what blocked deduplication.

This also fixes a bug when the NoDedup and MustDedup annotation pointed
at the same module.  We didn't record the presence of the module when it
was marked NoDedup which caused the MustDedup annotation to claim that
the referenced module did not exist.