[FIRRTL] Add module scope driver utils, NFC
Generalize a utility inside the RemoveInvalid pass for determining if a
value is driven by an invalid value within module scope into two FIRRTL
dialect utilities:

  1. getModuleScopedDriver
  2. isModuleScopedDrivenBy

(1) will work backwards from a value to find its driving value.  This
will look through wires and nodes based on two boolean parameters the
user provides.  This pedantic choice of separating looking through wires
and nodes, respectfully, is done because of certain passes in the Scala
FIRRTL Compiler (SFC) which only look through wires, nodes, or wires and
nodes.  (2) is a wrapper around (1) that will tell the user if a value
is driven by a specific operand type.  This is intended to be used for
certain SFC-exactness checks that want some value to be driven by an
invalid value or a constant.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>