[FIRRTL] Add FMemModule operation
This adds a FMemModule operation, which is an external module used for
FIRRTL SRAMs.  It is almost exactly the same, except it doesn't support
FExtModule's parameterization, does not have  a `defname`, and it
contains required attributes for memory metadata. It was convenient to
not use FExtModules with a bunch of extra attributes stored on it to
make sure that the attributes were properly checked.

The set of attributes stored on the memory are those needed to create all
the metadata files.  The types used for the attributes were generally
the same as those used by the memory generator schema. It might make
sense to revisit the types and use signless integer types in the future.