[FIRRTL] Add AddSeqMemPorts pass
This adds a new FIRRTL transformation which looks for
AddSeqMemPortsAnnotation and attaches the extra port specified to all
SRAMs under the DUT.  The extra ports are always of type `UInt<>` with a
user specified width.  The extra ports are wired up through the top of
the DUT and tied off to 0 in the testharness.

This pass requires that memories have been lowered to modules, and will
emit an error if it finds a regular mem op in the module.  This is
because it is not possible to add arbitrary ports to a memop that don't
match the schema of a memory r/w/rw port.