[FIRRTL] Make combmem clock analysis see through subfields
When lowering combinational memories, it is important to correctly
detect which ports are using the same clock.  Before this change, the
analysis would compare which value was connected to the clock.  This
could become a problem when the clock used is a subfield of a bundle,
and each port is using a different value. The SFC MemtoRefOfVec
transform compares the expression tree feeding into the clock, which
includes subfield, but not statements like nodes.