[FIRRTL][FullAsyncResetTransform] Preseve synchronous reset value (#3078)
When the FullAsyncResetTransform replaces the reset of a register it
replaces all other connections to the register with a `mux(reset,
resetValue, originalValue)`.  This caused problems when the value was
connected inside of a When block. The old reset value would become gated
on the When op's condition.  The original reset value needed to have
higher priority than the original conditions. This change adds an
explicit self-connect to the register which is properly muxed with the
reset.

Input:
```firrtl
reg r : UInt<8>, clock with :
  (reset => (reset, UInt<8>(1)))

when cond:
  r <= x
```

Original output:
```firrtl
reg r : UInt<8>, clock with :
  (reset => (async, UInt<8>(0)))

when cond:
  r = mux (reset, UInt<8>(1), x)
```

New output:
```firrtl
reg r : UInt<8>, clock with :
  (reset => (async, UInt<8>(0)))

r <= mux(reset, UInt<8>(1), r)

when cond:
  r = mux (reset, UInt<8>(1), x)
```