[FIRRTL][LowerMemory] Create wrapper module
When the SFC lowers memories to ext modules, they also create a wrapper
module which just instantiates the external module.  MFC did not create
this wrapper module, since the extra level of abstraction did not seem
necessary. We have found that to be a drop in replacement for SFC, we
need these wrapper modules so that Module and Instance paths are all
identical. This commit adds wrapper modules following the style of the
SFC.

This commit also fixes how non-local annotations are handled by this
pass.  When the memory operation has a non-local annotation attached, we
need to move the annotation over to the new module, update the NLA to
point to the module instead of the instance, and put an NLA reference on
the instance.