[Calyx] Add group-invariant code motion pass
The assumption here is that any operation which is not a special group operation (`calyx.group_done,calyx.group_go,calyx.assign`) is some form of combinational computation that can be lifted to the wire scope.

The pattern is added as a separate pass, but could be added as a canonicalization of `GroupOp`s.
I'd be in favor of adding this as a canonicalization of `GroupOp`s. GICM does not mean that we suddenly get unwanted sharing of `comb` operation results (e.g. as seen in the test, an identical `comb` operation is emitted twice in the wire scope - CSE is not part of `comb` canonicalizers, which i presume is due to fan-out reasons). However, what it does enable is easy detection of identical `comb` operations in the `wires` region which may be shared. The only downside to GICM in a canonicalizer is, that it might obscure the IR a slight bit, since `comb` operations are no longer inside the groups where they were initially placed during either human IR writing, or when some other pass generated the IR.
Let me know what you think.